Browser naming refactor
