+++ lib/feature/browser_v2/screens/main/browser_main_screen_wm.dart
@@ -11,11 +11,11 @@ import 'package:app/feature/browser_v2/screens/main/browser_main_screen.dart';
 import 'package:app/feature/browser_v2/screens/main/browser_main_screen_model.dart';
 import 'package:app/feature/browser_v2/screens/main/data/browser_render_manager.dart';
 import 'package:app/feature/browser_v2/screens/main/data/menu_data.dart';
-import 'package:app/feature/browser_v2/screens/main/helpers/menu_animation_helper.dart';
+import 'package:app/feature/browser_v2/screens/main/menu_animation_helper.dart';
+import 'package:app/feature/browser_v2/screens/main/widgets/control_panels/navigation_panel/url_action_sheet.dart';
 import 'package:app/feature/browser_v2/screens/main/widgets/tab_animated_view/tab_animation_type.dart';
 import 'package:app/feature/browser_v2/screens/main/widgets/tab_menu/data.dart';
 import 'package:app/feature/browser_v2/screens/main/widgets/tab_menu/tab_menu.dart';
-import 'package:app/feature/browser_v2/screens/main/widgets/url_menu.dart';
 import 'package:app/utils/clipboard_utils.dart';
 import 'package:app/utils/focus_utils.dart';
 import 'package:elementary/elementary.dart';
@@ -280,18 +280,18 @@ class BrowserMainScreenWidgetModel
   }
 
   Future<void> onPressedCurrentUrlMenu(String tabId) async {
-    final result = await showUrlMenu(context);
+    final result = await showUrlActionSheet(context);
 
     if (result == null) {
       return;
     }
 
     switch (result) {
-      case BrowserUrlMenuValue.copyUrl:
+      case BrowserUrlActionValue.copyUrl:
         model.copyTabUrl(tabId);
-      case BrowserUrlMenuValue.addBookmark:
+      case BrowserUrlActionValue.addBookmark:
         model.addUrlToBookmark(tabId);
-      case BrowserUrlMenuValue.clearFromHistory:
+      case BrowserUrlActionValue.clearFromHistory:
         model.clearUrlFromHistory(tabId);
     }
   }
