--- lib/feature/browser_v2/screens/main/data/control_panels_data.dart
@@ -1,10 +1,0 @@
-class BrowserMainControlPanelData {
-  BrowserMainControlPanelData({
-    bool? isCanGoBack,
-    bool? isCanGoForward,
-  })  : isCanGoBack = isCanGoBack ?? false,
-        isCanGoForward = isCanGoForward ?? false;
-
-  final bool isCanGoBack;
-  final bool isCanGoForward;
-}
