+++ lib/feature/browser_v2/screens/main/data/toolbar_data.dart
@@ -1,0 +1,10 @@
+class ToolbarData {
+  ToolbarData({
+    bool? isCanGoBack,
+    bool? isCanGoForward,
+  })  : isCanGoBack = isCanGoBack ?? false,
+        isCanGoForward = isCanGoForward ?? false;
+
+  final bool isCanGoBack;
+  final bool isCanGoForward;
+}
