+++ lib/feature/browser_v2/screens/main/widgets/browser_progress_indicator.dart
@@ -1,6 +1,6 @@
 import 'package:app/feature/browser_v2/screens/main/data/menu_data.dart';
-import 'package:app/feature/browser_v2/screens/main/widgets/control_panel/page_control_panel/page_control_panel.dart';
-import 'package:app/feature/browser_v2/screens/main/widgets/url_text_panel.dart';
+import 'package:app/feature/browser_v2/screens/main/widgets/control_panels/page_control_panel.dart';
+import 'package:app/feature/browser_v2/screens/main/widgets/control_panels/navigation_panel/host_panel.dart';
 import 'package:elementary_helper/elementary_helper.dart';
 import 'package:flutter/material.dart';
 import 'package:ui_components_lib/v2/ui_components_lib_v2.dart';
@@ -36,7 +36,7 @@ class _BrowserProgressIndicatorState extends State<BrowserProgressIndicator> {
             0,
             switch (type) {
               MenuType.view => -BrowserPageControlPanel.minHeight,
-              MenuType.url => -UrlTextPanel.height,
+              MenuType.url => -HostPanel.height,
               _ => 500,
             },
           ),
