+++ lib/feature/browser_v2/screens/main/widgets/control_panels/background_blur.dart
