+++ lib/feature/browser_v2/screens/main/widgets/pages/page/helpers/events_helper.dart
